!!![BUGFIX] Fix for sql_mode=only_full_group_by
In newer MySQL versions the query in GlossaryRepository::findAllForIndex()
will fail if sql_mode contains only_full_group_by.

The problem is the ORDER BY clause which does not match the GROUP BY.

This patch replaces the Extbase query entirely with Doctrine-dbal
functionality.

The use of language fetching via getAspect() will not work for TYPO3
previous to 9.5

Resolves: #10